Factor out an assertion builder utility for the rest of the assertion reworking.
