+++ tests/helpers/custom-helpers/assert-github-user-ok.js
@@ -1,35 +1,21 @@
 import QUnit from 'qunit';
 import Ember from 'ember';
+import assertionBuilder from '../utils/defined-attribute-assertion-builder';
 
-QUnit.assert.githubUserOk = function(user, message) {
-  let allDefined = true;
-  let undefinedKeys = [];
-  [
-    'id',
-    'login',
-    'name',
-    'type',
-    'avatarUrl',
-    'publicRepos',
-    'publicGists',
-    'followers',
-    'following',
-    'createdAt',
-    'updatedAt',
-    'url'
-  ].forEach(key => {
-    if( user.get(key) === undefined ) {
-      allDefined = false;
-      undefinedKeys.push(key);
-    }
-  });
-  this.pushResult({
-    result: allDefined,
-    actual: allDefined,
-    expected: true,
-    message: `${message}: ${undefinedKeys.join(',')}`
-  });
-};
+QUnit.assert.githubUserOk = assertionBuilder([
+  'id',
+  'login',
+  'name',
+  'type',
+  'avatarUrl',
+  'publicRepos',
+  'publicGists',
+  'followers',
+  'following',
+  'createdAt',
+  'updatedAt',
+  'url'
+]);
 
 export default Ember.Test.registerHelper(
   'assertGithubUserOk',
